move set_random_seed into conversion mixin
