+++ autoemulate/experimental/compare.py
@@ -1,10 +1,8 @@
 import logging
-import random
 import warnings
 from typing import Any
 
 import numpy as np
-import torch
 from sklearn.model_selection import BaseCrossValidator, KFold
 
 from autoemulate.experimental.data.utils import ConversionMixin
@@ -103,21 +101,3 @@ class AutoEmulate(ConversionMixin, TorchDeviceMixin):
             }
             self.log_compare(model_cls, best_model_config, r2_score, rmse_score)
         return models_evaluated
-
-    def set_random_seed(self, seed: int, deterministic: bool = False):
-        """Set random seed for Python, NumPy and PyTorch.
-
-        Parameters
-        ----------
-        seed : int
-            The random seed to use.
-        deterministic : bool
-            Use "deterministic" algorithms in PyTorch.
-        """
-        random.seed(seed)
-        np.random.seed(seed)
-        torch.manual_seed(seed)
-        torch.cuda.manual_seed(seed)
-        if deterministic:
-            torch.backends.cudnn.benchmark = False
-            torch.use_deterministic_algorithms(True)
